Restructure the basepaltable subsystem to support up to 256 palettes and use dynamic allocation like palookup and blendtable.
git-svn-id: https://svn.eduke32.com/eduke32@5348 1a8010ca-5511-0410-912e-c29ae57300e0

+++ polymer/eduke32/source/game.h
@@ -378,7 +378,7 @@ void G_UpdatePlayerFromMenu(void);
 void M32RunScript(const char *s);
 void P_DoQuote(int32_t q,DukePlayer_t *p);
 extern int32_t textsc(int32_t sc);
-void P_SetGamePalette(DukePlayer_t *player,uint8_t palid,int32_t set);
+void P_SetGamePalette(DukePlayer_t *player, uint32_t palid, int32_t set);
 
 extern int32_t G_GetStringLineLength(const char *text, const char *end, const int32_t iter);
 extern int32_t G_GetStringNumLines(const char *text, const char *end, const int32_t iter);
